seedengine: move to the root internal package
